% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifactor.R
\name{analyize.stepwise}
\alias{analyize.stepwise}
\title{Performs stepwise CAPA analysis on EEG data, saving
result plots for each step and writing a .csv data file
with every epoch/subepoch pair containing an anomaly.}
\usage{
analyize.stepwise(eeg, step_size, res, alpha = 8, beta = 1, thresh = 3)
}
\arguments{
\item{eeg}{An eeg object.}

\item{step_size}{Size in seconds of the sequence of data analyized
on each step.}

\item{res}{Resolution at which to perform the analyses}

\item{alpha}{Threshold of strength significance for collective anomalies}

\item{beta}{Threshold of strength significance for point anomalies}

\item{thresh}{How many seconds anomaly n must be from anomaly (n - 1) to
consider them part of a same cluster?}
}
\description{
Performs stepwise CAPA analysis on EEG data, saving
result plots for each step and writing a .csv data file
with every epoch/subepoch pair containing an anomaly.
}
