% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\docType{class}
\name{analysis-class}
\alias{analysis-class}
\title{Analysis class.}
\description{
Analysis class.
}
\section{Slots}{

\describe{
\item{\code{canoms}}{A collective anomalies data frame.}

\item{\code{signals}}{A point anomalies data frame.}

\item{\code{The}}{eeg upon which analysis is conducted.}
}}

