% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifactor.R
\name{analyze}
\alias{analyze}
\title{Perform M-CAPA analysis on EEG data in a given range of seconds
and return results filtered by anomaly strength.}
\usage{
analyze(eeg, s, e, res = 1, alpha = 8, beta = 1, thresh = 3, time = TRUE)
}
\arguments{
\item{eeg}{An eeg object}

\item{s}{int First second of timespan to analyze}

\item{e}{int Last second of timespan to analyze}

\item{res}{int Resolution at which to perform analysis}

\item{alpha}{float Threshold of strength significance for collective anomalies}

\item{beta}{float Threshold of strength significance for point anomalies}

\item{thresh}{int How many seconds collective anomaly n must be from
collective anomaly (n - 1) to consider them part of a same cluster?}

\item{time}{bool Show process time?}
}
\value{
An analysis object
}
\description{
Perform M-CAPA analysis on EEG data in a given range of seconds
and return results filtered by anomaly strength.
}
