% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{artf_reject,eeg-method}
\alias{artf_reject,eeg-method}
\title{Given an eeg object and an analysis object,
returns an artifact-rejected version of the
eeg.}
\usage{
\S4method{artf_reject}{eeg}(object, analysis)
}
\arguments{
\item{object}{An eeg object.}

\item{analysis}{An analysis object}
}
\value{
An EEG object.
}
\description{
Given an eeg object and an analysis object,
returns an artifact-rejected version of the
eeg.
}
