% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{bandpass,eeg-method}
\alias{bandpass,eeg-method}
\title{Given an eeg object and a numeric frequency n,
applies a bandpass filter.}
\usage{
\S4method{bandpass}{eeg}(object, l, h)
}
\arguments{
\item{object}{An eeg object.}

\item{l}{lower bound in Hz}

\item{h}{higher bound in Hz}
}
\value{
A new filtered EEG object
}
\description{
Given an eeg object and a numeric frequency n,
applies a bandpass filter.
}
