% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{create.epoch.data}
\alias{create.epoch.data}
\title{Create an empty data frame to be filled with epoch-subepoch
anomalous pairs during stepwise analysis.}
\usage{
create.epoch.data()
}
\value{
An empty dataframe.
}
\description{
Create an empty data frame to be filled with epoch-subepoch
anomalous pairs during stepwise analysis.
}
