% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dman.R
\name{create_epoch_data}
\alias{create_epoch_data}
\title{Create an empty data frame to be filled with epoch-subepoch
anomalous pairs during stepwise analysis.}
\usage{
create_epoch_data()
}
\value{
An empty dataframe.
}
\description{
Create an empty data frame to be filled with epoch-subepoch
anomalous pairs during stepwise analysis.
}
