% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cln.R
\name{detect_clusters}
\alias{detect_clusters}
\title{Given a single-channel canoms data frame, returns a list of all
 clusters of anomalies. Each element in the list is a pair of
the form (s, e) containing the start and the end indexes of each cluster.}
\usage{
detect_clusters(canoms, cluster_threshold)
}
\arguments{
\item{canoms}{A canoms data frame with a single channel.}
}
\value{
A list of (s, e) pairs.
}
\description{
Given a single-channel canoms data frame, returns a list of all
 clusters of anomalies. Each element in the list is a pair of
the form (s, e) containing the start and the end indexes of each cluster.
}
