% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{drop_subepochs,eeg-method}
\alias{drop_subepochs,eeg-method}
\title{Given an eeg object, a list of epochs and
a list of subepochs, removes epoch-subepoch
pairs from the eeg data.}
\usage{
\S4method{drop_subepochs}{eeg}(object, epochs, subepochs, epoch = 30)
}
\arguments{
\item{object}{An eeg object.}

\item{epoch}{An natural number}
}
\value{
An EEG object.
}
\description{
Note that elements in the epoch and subepoch
lists are assumed to have an element-wise
association. This means that, if this lists
are (e_1, ..., e_n) and (s_1, ..., s_n), then
epoch/subepoch pairs (e_i, s_i) are removed
for i in [1, n].
}
