% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\docType{class}
\name{eeg-class}
\alias{eeg-class}
\title{EEG class.}
\description{
EEG class.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame containing EEG records}

\item{\code{signals}}{A data frame containing signal information (may be empty).}
}}

