% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{class}
\name{EEG}
\alias{EEG}
\title{Class providing object with methods for EEG analysis, visualization, and manipulation.}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} with methods for EEG analysis,
visualization and manipulation.
}
\description{
Class providing object with methods for EEG analysis, visualization, and manipulation.

Class providing object with methods for EEG analysis, visualization, and manipulation.
}
\section{Methods}{

\describe{
  \item{\code{subset_by_seconds(s, e)}}{Subsets the EEG from second s to
  second e. Both \code{s} and \code{e} must be numeric values that exist in the `Time`
  column of the EEG data. Modification is performed inplace.}
  \item{\code{subset(s, e)}}{Subsets the EEG from epoch s to epoch e. The
  interval is inclusive. Modification is performed inplace.}
  \item{\code{resample(n)}}{Reduces the EEG data by removing on every
  \code{n} samples. This is a (very) brute resampling method and should only
  be used for the purpose of accelerating certain analyses, such as
  artifact detection.}
  \item{\code{low_pass(n)}}{Applies a low-pass Butterworth filter with a
  filter frequency \code{n} (in Hz).}
  \item{\code{high_pass(n)}}{Applies a high-pass Butterworth filter with a
  filter frequency \code{n} (in Hz).}
  \item{\code{bandpass(l, h)}}{Applies a bandpass Butterworth filter with
  filter frequencies \code{l} (lower bound) and \code{h} (upper bound) (in Hz).}
  \item{\code{plot_channel(channel)}}{Plots the ith signal in data, where i =
  \code{channel}}.
  \item{\code{plot}}{Plots the EEG data.}
  \item{\code{iplot}}{Draws an interactive plot of the EEG data.}
  \item{\code{plot}}{Plots the EEG data.}
  \item{\code{drop_epochs(epochs, epoch = 30)}}{Given a list `epochs` of
  integer values, removes those epochs from the EEG data. The argument
  `epoch` determins how many seconds are understood to comprise an epoch and
  defaults to 30.}
  \item{\code{drop_subepochs(epochs, subepochs, epoch = 30)}}{Given numeric
  lists of \code{subepochs} and \code{epochs}, removes epoch/subepoch pairs from the
  EEG data. Elements in both lists are assumed to have an element-wise
  association, meaning that if \code{(e_1, ..., e_n), (s_1, ..., s_n)} are
  both lists, then epoch/subepoch pairs \code{(e_i, s_i)} are removed. The
  argument `epoch` is an integer determining how many seconds are understood
  to comprise an epoch (defaults to 30).}
  \item{\code{artf(alpha = 8, beta = 1)}}{Performs direct artifact detection
  over the EEG data. \code{alpha, beta} are numeric significance thresholds
  for collective and point anomalies respectively.}
  \item{\code{artf_stepwise(step_size = 30, alpha=8)}}{Performs stepwise
  artifact detection over the EEG data, by steps of \code{step_size} seconds.
  \code{alpha} is the significance threshold for collective anomalies.}
  \item{\code{get_contaminated_channels}}{Returns a vector with all EEG
  channels found to contain artifacts.}
  \item{\code{plot_artifacts(size = 0.2)}}{Plots EEG data with artifacts
  represented by red points of a certain \code{size}.}
  \item{\code{sfilter(x, f = minmax_normalization)}}{After normalizing
  artifact strength with a normalization function \code{f}, filters out all
  artifacts of strength lesser than \code{x}.}
  \item{\code{artf_reject()}}{Returns an artifact rejected copy of this EEG.}
  \item{\code{spectrogram(channel, max_freq = 30, freq = 4)}}{Plots the
  spectrogram of a given \code{channel}. \code{freq} is the resolution of the
  plot.}
  \item{\code{compute_psd()}}{Computes the power spectrum of this EEG and
  sotres it on the psd field.}
  \item{\code{plot_psd(xlim = 250)}}{Plots the PSD of this EEG.}
  \item{\code{iplot_psd(xlim)}}{Produces an interactive plot of the PSD.}}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Stores the EEG data in a data frame format (tibble by default).}

\item{\code{signals}}{Stores a data frame with information on the EEG signals, if a
`signals` file was provided on initialization.}

\item{\code{canoms}}{A data frame with all data pertaining to collective anomalies
(artifacts) found on the EEG. Empty by default and until artifact analysis is
performed.}

\item{\code{panoms}}{A data frame with all data pertaining to point anomalies
(artifacts) found on the EEG. Empty by default and until artifact analysis is
performed.}

\item{\code{psd}}{Power spectrum density data of this EEG. Empty by default and
until PSD is computed on the EEG.}

\item{\code{fs}}{Sampling frequency of the EEG.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EEG-new}{\code{EEG$new()}}
\item \href{#method-EEG-subset_by_seconds}{\code{EEG$subset_by_seconds()}}
\item \href{#method-EEG-subset}{\code{EEG$subset()}}
\item \href{#method-EEG-resample}{\code{EEG$resample()}}
\item \href{#method-EEG-get_fs}{\code{EEG$get_fs()}}
\item \href{#method-EEG-vlow_pass}{\code{EEG$vlow_pass()}}
\item \href{#method-EEG-vhigh_pass}{\code{EEG$vhigh_pass()}}
\item \href{#method-EEG-vbandpass}{\code{EEG$vbandpass()}}
\item \href{#method-EEG-low_pass}{\code{EEG$low_pass()}}
\item \href{#method-EEG-high_pass}{\code{EEG$high_pass()}}
\item \href{#method-EEG-bandpass}{\code{EEG$bandpass()}}
\item \href{#method-EEG-plot_channel}{\code{EEG$plot_channel()}}
\item \href{#method-EEG-plot}{\code{EEG$plot()}}
\item \href{#method-EEG-iplot}{\code{EEG$iplot()}}
\item \href{#method-EEG-drop_epochs}{\code{EEG$drop_epochs()}}
\item \href{#method-EEG-drop_subepochs}{\code{EEG$drop_subepochs()}}
\item \href{#method-EEG-artf}{\code{EEG$artf()}}
\item \href{#method-EEG-artf_stepwise}{\code{EEG$artf_stepwise()}}
\item \href{#method-EEG-get_contaminated_channels}{\code{EEG$get_contaminated_channels()}}
\item \href{#method-EEG-set_plot_data}{\code{EEG$set_plot_data()}}
\item \href{#method-EEG-plot_channel_artifacts}{\code{EEG$plot_channel_artifacts()}}
\item \href{#method-EEG-plot_artifacts}{\code{EEG$plot_artifacts()}}
\item \href{#method-EEG-sfilter}{\code{EEG$sfilter()}}
\item \href{#method-EEG-get_contaminated_epochs}{\code{EEG$get_contaminated_epochs()}}
\item \href{#method-EEG-artf_reject}{\code{EEG$artf_reject()}}
\item \href{#method-EEG-spectrogram}{\code{EEG$spectrogram()}}
\item \href{#method-EEG-channel_psd}{\code{EEG$channel_psd()}}
\item \href{#method-EEG-compute_psd}{\code{EEG$compute_psd()}}
\item \href{#method-EEG-plot_psd}{\code{EEG$plot_psd()}}
\item \href{#method-EEG-iplot_psd}{\code{EEG$iplot_psd()}}
\item \href{#method-EEG-clone}{\code{EEG$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-new"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$new(data_file, signals_file = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-subset_by_seconds"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-subset_by_seconds}{}}}
\subsection{Method \code{subset_by_seconds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$subset_by_seconds(s, e)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-subset"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-subset}{}}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$subset(s, e, epoch = 30)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-resample"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-resample}{}}}
\subsection{Method \code{resample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$resample(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-get_fs"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-get_fs}{}}}
\subsection{Method \code{get_fs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$get_fs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-vlow_pass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-vlow_pass}{}}}
\subsection{Method \code{vlow_pass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$vlow_pass(vec, n, fs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-vhigh_pass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-vhigh_pass}{}}}
\subsection{Method \code{vhigh_pass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$vhigh_pass(vec, n, fs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-vbandpass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-vbandpass}{}}}
\subsection{Method \code{vbandpass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$vbandpass(vec, l, h, fs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-low_pass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-low_pass}{}}}
\subsection{Method \code{low_pass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$low_pass(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-high_pass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-high_pass}{}}}
\subsection{Method \code{high_pass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$high_pass(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-bandpass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-bandpass}{}}}
\subsection{Method \code{bandpass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$bandpass(l, h)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-plot_channel"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-plot_channel}{}}}
\subsection{Method \code{plot_channel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$plot_channel(channel)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-plot"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-iplot"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-iplot}{}}}
\subsection{Method \code{iplot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$iplot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-drop_epochs"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-drop_epochs}{}}}
\subsection{Method \code{drop_epochs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$drop_epochs(epochs, epoch = 30)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-drop_subepochs"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-drop_subepochs}{}}}
\subsection{Method \code{drop_subepochs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$drop_subepochs(epochs, subepochs, epoch = 30)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-artf"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-artf}{}}}
\subsection{Method \code{artf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$artf(alpha = 8, beta = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-artf_stepwise"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-artf_stepwise}{}}}
\subsection{Method \code{artf_stepwise()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$artf_stepwise(step_size = 30, alpha = 8)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-get_contaminated_channels"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-get_contaminated_channels}{}}}
\subsection{Method \code{get_contaminated_channels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$get_contaminated_channels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-set_plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-set_plot_data}{}}}
\subsection{Method \code{set_plot_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$set_plot_data(chan)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-plot_channel_artifacts"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-plot_channel_artifacts}{}}}
\subsection{Method \code{plot_channel_artifacts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$plot_channel_artifacts(chan, size = 0.2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-plot_artifacts"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-plot_artifacts}{}}}
\subsection{Method \code{plot_artifacts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$plot_artifacts(size = 0.2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-sfilter"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-sfilter}{}}}
\subsection{Method \code{sfilter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$sfilter(x, f = minmax_normalization)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-get_contaminated_epochs"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-get_contaminated_epochs}{}}}
\subsection{Method \code{get_contaminated_epochs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$get_contaminated_epochs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-artf_reject"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-artf_reject}{}}}
\subsection{Method \code{artf_reject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$artf_reject()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-spectrogram"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-spectrogram}{}}}
\subsection{Method \code{spectrogram()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$spectrogram(channel, max_freq = 30, freq = 4)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-channel_psd"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-channel_psd}{}}}
\subsection{Method \code{channel_psd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$channel_psd(channel)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-compute_psd"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-compute_psd}{}}}
\subsection{Method \code{compute_psd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$compute_psd()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-plot_psd"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-plot_psd}{}}}
\subsection{Method \code{plot_psd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$plot_psd(xlim = 250)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-iplot_psd"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-iplot_psd}{}}}
\subsection{Method \code{iplot_psd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$iplot_psd(xlim = 250)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
