% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dman.R
\name{extract_epochs,analysis-method}
\alias{extract_epochs,analysis-method}
\title{data attribute is such subset.}
\usage{
\S4method{extract_epochs}{analysis}(object)
}
\arguments{
\item{object}{An eeg object.}

\item{s}{Starting time of the subset in seconds.}

\item{e}{Ending time of the subset in seconds.}
}
\value{
A new eeg whose data is the subset ranging from
second s to e of the object's data attribute.
}
\description{
data attribute is such subset.
}
