% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cln.R
\name{format.collectives}
\alias{format.collectives}
\title{Given a multi-channel canoms data frame, returns a formatted version of the data frame
where all anomalies belonging to the same clusterwere reduced to a single
anomaly in every record channel.}
\usage{
\method{format}{collectives}(anom_df, cluster_thresh)
}
\arguments{
\item{canoms}{A canoms data frame.}

\item{clusters}{A list of (s, e) pairs containing start and end indexes of all clusters in canoms.}
}
\value{
A canoms data frame.
}
\description{
Given a multi-channel canoms data frame, returns a formatted version of the data frame
where all anomalies belonging to the same clusterwere reduced to a single
anomaly in every record channel.
}
