% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cln.R
\name{format_time}
\alias{format_time}
\title{Given a lubridate time object s, returns a character formatted in the
hours:minutes:seconds.}
\usage{
format_time(s)
}
\arguments{
\item{s}{A lubridate time object.}
}
\value{
A character.
}
\description{
Given a lubridate time object s, returns a character formatted in the
hours:minutes:seconds.
}
