% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cln.R
\name{get.time}
\alias{get.time}
\title{Given a canoms or panoms data frame, returns a vector containing
the lubridate period objects corresponding to the time of occurrence,
in the EEG record, of every anomaly of the data frame.}
\usage{
get.time(anoms, data)
}
\arguments{
\item{anoms}{A canoms or panoms data frame.}
}
\value{
A vector of lubridate periods
}
\description{
Given a canoms or panoms data frame, returns a vector containing
the lubridate period objects corresponding to the time of occurrence,
in the EEG record, of every anomaly of the data frame.
}
