% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cln.R
\name{get_anom_maxims}
\alias{get_anom_maxims}
\title{Returns the maximum of all mean changes that occurred in
an anomalous cluster.}
\usage{
get_anom_maxims(canoms, start_points, end_points)
}
\arguments{
\item{start_points}{Start indexes of all clusters in canoms}

\item{anom_df}{A canoms dataframe.}

\item{threshold}{End indexes of all clusters in anoms}
}
\value{
Vector containing maximum mean change of each anomaly in canoms
}
\description{
Returns the maximum of all mean changes that occurred in
an anomalous cluster.
}
