% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{get_anomalous_channels,analysis-method}
\alias{get_anomalous_channels,analysis-method}
\title{Getter function retrieving all EEG channels containing
anomalies in the Analysis object.}
\usage{
\S4method{get_anomalous_channels}{analysis}(object)
}
\arguments{
\item{object}{An Analysis object.}
}
\value{
A vector containing all anomalous channels in the analysis.
}
\description{
Getter function retrieving all EEG channels containing
anomalies in the Analysis object.
}
