% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{get_channel_psd,eeg-method}
\alias{get_channel_psd,eeg-method}
\title{Given an eeg object, determine the number of values
that make up epoch seconds.}
\usage{
\S4method{get_channel_psd}{eeg}(object, channel, method = "welch")
}
\arguments{
\item{object}{An eeg object.}

\item{epoch}{A time in seconds.}
}
\value{
An integer representing the number of values
that make up a time-frame of length epoch.
}
\description{
Given an eeg object, determine the number of values
that make up epoch seconds.
}
