% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{get_epoch,eeg-method}
\alias{get_epoch,eeg-method}
\title{Returns new eeg whose data is a subset of given eeg
on the given epoch. Epoch counts begin with 0.}
\usage{
\S4method{get_epoch}{eeg}(object, which_epoch, epoch = 30)
}
\arguments{
\item{object}{An eeg object.}

\item{which_epoch}{What epoch to subset.}

\item{epoch}{How many seconds make up an epoch? Defaults to 30.}
}
\value{
A new eeg whose data is the requested epoch.
}
\description{
This function wraps a specific call of subset_eeg.
}
