% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{get_samples_in_epoch,eeg-method}
\alias{get_samples_in_epoch,eeg-method}
\title{Given an eeg object, determine the number of values
that make up epoch seconds.}
\usage{
\S4method{get_samples_in_epoch}{eeg}(object, epoch)
}
\arguments{
\item{object}{An eeg object.}

\item{epoch}{A time in seconds.}
}
\value{
An integer representing the number of values
that make up a time-frame of length epoch.
}
\description{
Given an eeg object, determine the number of values
that make up epoch seconds.
}
