% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cln.R
\name{get_time}
\alias{get_time}
\title{Given a canoms or panoms data frame, returns a vector containing
the time of occurrence of each anomaly on the EEG record.}
\usage{
get_time(anoms, data)
}
\arguments{
\item{anoms}{A canoms or panoms data frame.}
}
\value{
A numeric vector
}
\description{
Given a canoms or panoms data frame, returns a vector containing
the time of occurrence of each anomaly on the EEG record.
}
