% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{has_anoms,analysis-method}
\alias{has_anoms,analysis-method}
\title{Check if either canoms or panoms are non-empty data frames.}
\usage{
\S4method{has_anoms}{analysis}(object)
}
\arguments{
\item{object}{An Analysis object.}
}
\value{
bool
}
\description{
Check if either canoms or panoms are non-empty data frames.
}
