% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{high_pass,eeg-method}
\alias{high_pass,eeg-method}
\title{Given an eeg object and a numeric frequency n,
applies a low-pass Butterworth filter.}
\usage{
\S4method{high_pass}{eeg}(object, n)
}
\arguments{
\item{object}{An eeg object.}

\item{n}{Filter frequency in Hz}
}
\value{
A new filtered EEG object
}
\description{
Given an eeg object and a numeric frequency n,
applies a low-pass Butterworth filter.
}
