% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{iplot,eeg-method}
\alias{iplot,eeg-method}
\title{Given an EEG returns its interactive visualization.}
\usage{
\S4method{iplot}{eeg}(eeg)
}
\arguments{
\item{object}{An eeg object.}
}
\value{
A plotly figure.
}
\description{
Given an EEG returns its interactive visualization.
}
