% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{iplot.analysis,analysis-method}
\alias{iplot.analysis,analysis-method}
\title{Creates an interactive plotly plot depicting the analysis.
Since interactive analysis plotting may be expensive, the
plotly figure may be saved into an .html file for re-use.}
\usage{
\S4method{iplot.analysis}{analysis}(
  object,
  data = NULL,
  s,
  e,
  n_resample = as.integer(2),
  save = TRUE,
  show = FALSE
)
}
\arguments{
\item{object}{An Analysis object.}

\item{data}{(Optional) A data frame of format specified by the
set.iplot.data method, or a string with the path to a .csv stored
representation of such data frame. If none is provided, the iplot
data is created on method call.}

\item{s}{Initial second of the plot.}

\item{e}{Final second of the plot.}

\item{n_resample}{Integer indicating downsampling decimation.
If n is given, only one every n values are plotted. This should
be set to higher values (5 or 10) for very large plots (for example,
plots depicting a whole hour or more). Defaults to two.}

\item{save}{Save .html plot?}

\item{show}{Show the plot in browser?

format or not.}
}
\value{
A plot_grid object.
}
\description{
Creates an interactive plotly plot depicting the analysis.
Since interactive analysis plotting may be expensive, the
plotly figure may be saved into an .html file for re-use.
}
