% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{iplot.eeg,eeg-method}
\alias{iplot.eeg,eeg-method}
\title{Given an eeg object, draw an interactive visualization
of all channels.}
\usage{
\S4method{iplot.eeg}{eeg}(object, s = -1, e = -1, joint = FALSE, save = TRUE, show = FALSE)
}
\arguments{
\item{object}{Immediately show the plot after creating it? Defaults
to False.}

\item{s}{Starting second of the plot. Defaults to -1, which
means the starting second of the plot will be that of the EEG.}

\item{e}{Last second of the plot. Defaults to -1, which
means the last second of the plot will be that of the EEG.}

\item{joint}{Should channels be drawn one on top of the other?
Defaults to false.}
}
\description{
Given an eeg object, draw an interactive visualization
of all channels.
}
