% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{iplot_psd}
\alias{iplot_psd}
\title{Given an PSD data frame as returned by the psd(eeg, ...) function,
create an interactive plot of the spectrum of all channels.}
\usage{
iplot_psd(psd)
}
\arguments{
\item{object}{An data frame as returned by the psd(eeg, ...) function.}
}
\value{
A plotly figure.
}
\description{
Given an PSD data frame as returned by the psd(eeg, ...) function,
create an interactive plot of the spectrum of all channels.
}
