% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cln.R
\name{is_clustered}
\alias{is_clustered}
\title{Given the ocurrence time of an anomaly, the ocurrence time
of the previous anomaly and a certain threshold, determine
if both anomalies can be considered to be in the same cluster.}
\usage{
is_clustered(previous_anom_time, time, threshold)
}
\arguments{
\item{previous_anom_time}{Formated time of (n-1)th anomaly.}

\item{time}{Formated time of nth anomaly.}

\item{threshold}{Integer determining how many seconds away two
anomalies must be to not be considered part of the same cluster.}
}
\value{
bool
}
\description{
Given the ocurrence time of an anomaly, the ocurrence time
of the previous anomaly and a certain threshold, determine
if both anomalies can be considered to be in the same cluster.
}
