% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cln.R
\name{join.clusters}
\alias{join.clusters}
\title{Given a single-variate canoms data frame, returns an equal size
or smaller data frame where all anomalies belonging to the same
cluster were reduced to a single anomaly.}
\usage{
join.clusters(canoms, clusters)
}
\arguments{
\item{canoms}{A canoms data frame with a single channel.}

\item{clusters}{A list of (s, e) pairs containing start and end
indexes of all clusters in canoms.}
}
\value{
A canoms data frame.
}
\description{
Given a single-variate canoms data frame, returns an equal size
or smaller data frame where all anomalies belonging to the same
cluster were reduced to a single anomaly.
}
