% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{load_eeg}
\alias{load_eeg}
\title{Read a .csv data file containing EEG data and an optionall}
\usage{
load_eeg(data_file, signals_file = NULL)
}
\arguments{
\item{data_file}{.csv file containing eeg data.}

\item{signals_file}{.csv file containing signal information}
}
\value{
An eeg object.
}
\description{
signals file and return an eeg object. If the signals file is provided,
 channel names are appropriately set.
}
