% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{low_pass,eeg-method}
\alias{low_pass,eeg-method}
\title{Given an eeg object and a numeric frequency n,
applies a 1/n Hz low-pass Butterworth filter.}
\usage{
\S4method{low_pass}{eeg}(object, n)
}
\arguments{
\item{object}{An eeg object.}

\item{n}{Filter frequency.}
}
\value{
A new filtered EEG object
}
\description{
Given an eeg object and a numeric frequency n,
applies a 1/n Hz low-pass Butterworth filter.
}
