% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{merge,analysis-method}
\alias{merge,analysis-method}
\title{Merges two analysis into one}
\usage{
\S4method{merge}{analysis}(object, an)
}
\arguments{
\item{object}{An analysis object}

\item{an}{The analysis object being merged into the first}
}
\value{
An analysis object.
}
\description{
Merges two analysis into one
}
