% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{na.omit,eeg-method}
\alias{na.omit,eeg-method}
\title{NA omit for the EEG class that applies na.omit() to the data slot.}
\usage{
\S4method{na.omit}{eeg}(object)
}
\arguments{
\item{object}{An EEG object.}
}
\description{
NA omit for the EEG class that applies na.omit() to the data slot.
}
