% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{partition.eeg,eeg-method}
\alias{partition.eeg,eeg-method}
\title{Given an eeg object, determine the subset of
its data attribute that contains records from
second s to e, and return a new eeg object whose
data attribute is such subset.}
\usage{
\S4method{partition.eeg}{eeg}(object, s, e)
}
\arguments{
\item{object}{An eeg object.}

\item{s}{Starting time of the subset in seconds.}

\item{e}{Ending time of the subset in seconds.}
}
\value{
A new eeg whose data is the subset ranging from
second s to e of the object's data attribute.
}
\description{
Given an eeg object, determine the subset of
its data attribute that contains records from
second s to e, and return a new eeg object whose
data attribute is such subset.
}
