% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{plot.analysis,analysis-method}
\alias{plot.analysis,analysis-method}
\title{Wrapper and standard analysis plotting function.
Plots all analyzed channels with the option of saving
the plot.}
\usage{
\S4method{plot.analysis}{analysis}(object, save = FALSE)
}
\arguments{
\item{object}{An Analysis object.}

\item{save}{A bool declaring whether to save the image in .png format or not.}
}
\value{
A plot_grid object.
}
\description{
Wrapper and standard analysis plotting function.
Plots all analyzed channels with the option of saving
the plot.
}
