% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{plot.channel,eeg-method}
\alias{plot.channel,eeg-method}
\title{Given an eeg object and a channel integer n,
plots EEG record of the nth channel.}
\usage{
\S4method{plot.channel}{eeg}(object, channel)
}
\arguments{
\item{object}{An eeg object.}

\item{channel}{An integer indicating index of channel to plot.}
}
\value{
A ggplot object.
}
\description{
Given an eeg object and a channel integer n,
plots EEG record of the nth channel.
}
