% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{plot.channel.analysis,analysis-method}
\alias{plot.channel.analysis,analysis-method}
\title{Wrapper function that plots an analyzed channel with
any kind of anomalies in it.}
\usage{
\S4method{plot.channel.analysis}{analysis}(object, channel)
}
\arguments{
\item{object}{An Analysis object.}

\item{channel}{An integer.}
}
\value{
A ggplot.
}
\description{
Wrapper function that plots an analyzed channel with
any kind of anomalies in it.
}
