% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{plot_analysis,analysis-method}
\alias{plot_analysis,analysis-method}
\title{Plots all channels with any kind of anomalies.}
\usage{
\S4method{plot_analysis}{analysis}(object, size = 1)
}
\arguments{
\item{object}{An Analysis object.}
}
\value{
A plot_grid object.
}
\description{
Plots all channels with any kind of anomalies.
}
