% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{plot_anoms,analysis-method}
\alias{plot_anoms,analysis-method}
\title{Plots both point and collective anomalies in one of the EEG channels.}
\usage{
\S4method{plot_anoms}{analysis}(object, channel)
}
\arguments{
\item{object}{An Analysis object.}

\item{channel}{An integer.}
}
\value{
A ggplot.
}
\description{
Plots both point and collective anomalies in one of the EEG channels.
}
