% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{plot_clusters,analysis-method}
\alias{plot_clusters,analysis-method}
\title{Plots only collective anomalies in one of the EEG channels.}
\usage{
\S4method{plot_clusters}{analysis}(object, channel)
}
\arguments{
\item{object}{An Analysis object.}

\item{channel}{An integer.}
}
\value{
A ggplot.
}
\description{
Plots only collective anomalies in one of the EEG channels.
}
