% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{plot_eeg,eeg-method}
\alias{plot_eeg,eeg-method}
\title{Given an eeg object, plot all the EEG channels
in a vertical single-column layout.}
\usage{
\S4method{plot_eeg}{eeg}(object)
}
\arguments{
\item{object}{An eeg object.}
}
\value{
A plot_grid object.
}
\description{
Given an eeg object, plot all the EEG channels
in a vertical single-column layout.
}
