% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{plot_points,analysis-method}
\alias{plot_points,analysis-method}
\title{Plots only point anomalies in one of the EEG channels.}
\usage{
\S4method{plot_points}{analysis}(object, channel)
}
\arguments{
\item{object}{An Analysis object.}

\item{channel}{An integer.}
}
\value{
A ggplot.
}
\description{
Plots only point anomalies in one of the EEG channels.
}
