% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{plot_psd}
\alias{plot_psd}
\title{Given an PSD data frame as returned by the psd(eeg, ...) function,
plot the spectrum of all channels.}
\usage{
plot_psd(psd, xlim = 250)
}
\arguments{
\item{object}{An data frame as returned by the psd(eeg, ...) function.}
}
\value{
A ggplot object.
}
\description{
Given an PSD data frame as returned by the psd(eeg, ...) function,
plot the spectrum of all channels.
}
