% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{remove_epoch,eeg-method}
\alias{remove_epoch,eeg-method}
\title{Given an eeg object and an epoch, removes
all samples from that epoch from the EEG data.}
\usage{
\S4method{remove_epoch}{eeg}(object, epoch)
}
\arguments{
\item{object}{An eeg object.}

\item{epoch}{An natural number}
}
\value{
An EEG object.
}
\description{
Notice that the extremes of the epoch are kept.
}
