% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{remove_spindles_out_of_range}
\alias{remove_spindles_out_of_range}
\title{Given a boolean vector x, sets to FALSE all sequences of TRUE values
with less than min or more than max elements.
In the context of spindle detection, the boolean vector is understood
signal which indexes of the EEG sample were found to contain spindles.}
\usage{
remove_spindles_out_of_range(x, min, max)
}
\description{
Given a boolean vector x, sets to FALSE all sequences of TRUE values
with less than min or more than max elements.
In the context of spindle detection, the boolean vector is understood
signal which indexes of the EEG sample were found to contain spindles.
}
