% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{resample_eeg,eeg-method}
\alias{resample_eeg,eeg-method}
\title{Given an eeg object and an integer n,
subset the EEG's data by keeping only one every
n values and return a new EEG with the subsetted data.}
\usage{
\S4method{resample_eeg}{eeg}(object, n)
}
\arguments{
\item{object}{An eeg object.}

\item{n}{An integer.}
}
\value{
A new filtered EEG object
}
\description{
Logically, the new EEG will have a length of 1/n
the length of the subsetted EEG.
}
