% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rms_spindle}
\alias{rms_spindle}
\title{Given a numeric vector x, returns a boolean vector whose ith truth value
indicates whether the ith element of x was above the 0.95 percentile of
rms(x).}
\usage{
rms_spindle(x, λ = 0.95)
}
\description{
Given a numeric vector x, returns a boolean vector whose ith truth value
indicates whether the ith element of x was above the 0.95 percentile of
rms(x).
}
