% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cln.R
\name{set.anom.epoch}
\alias{set.anom.epoch}
\title{Given a canoms or panoms data frame, returns an identical
canoms or panoms with the inclusion of epoch and subepoch columns.}
\usage{
set.anom.epoch(anoms)
}
\arguments{
\item{anoms}{A canoms or panoms data frame.}
}
\value{
A data frame
}
\description{
Given a canoms or panoms data frame, returns an identical
canoms or panoms with the inclusion of epoch and subepoch columns.
}
