% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{set.chan.iplot.data,analysis-method}
\alias{set.chan.iplot.data,analysis-method}
\title{Formats a channel's analysis results to a data frame
specifically designed for interactive plotting using
Python's plotly library.}
\usage{
\S4method{set.chan.iplot.data}{analysis}(object, chan)
}
\arguments{
\item{object}{An Analysis object.}

\item{chan}{The channel to be formatted for intercative plotting.}
}
\value{
A tibble.
}
\description{
Formats a channel's analysis results to a data frame
specifically designed for interactive plotting using
Python's plotly library.
}
