% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cln.R
\name{set.timevars}
\alias{set.timevars}
\title{Given a canoms or panoms data frame and its origin, returns an identical
canoms or panoms data frame with the inclusion of all time variables:
Time, Epoch and Subepoch.}
\usage{
set.timevars(anoms, data)
}
\arguments{
\item{anoms}{A canoms or panoms data frame.}
}
\value{
A data frame
}
\description{
Given a canoms or panoms data frame and its origin, returns an identical
canoms or panoms data frame with the inclusion of all time variables:
Time, Epoch and Subepoch.
}
