% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cln.R
\name{set_anom_epoch}
\alias{set_anom_epoch}
\title{Given a canoms or panoms data frame, returns an identical
canoms or panoms with the inclusion of epoch and subepoch columns.}
\usage{
set_anom_epoch(anoms)
}
\arguments{
\item{anoms}{A canoms or panoms data frame.}
}
\value{
A data frame
}
\description{
Given a canoms or panoms data frame, returns an identical
canoms or panoms with the inclusion of epoch and subepoch columns.
}
