% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{set_for_iplot,analysis-method}
\alias{set_for_iplot,analysis-method}
\title{Sets the necessary data frame for interactive analysis plotting using
Python's plotly library.}
\usage{
\S4method{set_for_iplot}{analysis}(object, save = FALSE)
}
\arguments{
\item{object}{An Analysis object.}

\item{save}{A bool declaring whether to save the resulting
data frame in in .csv format or not.}
}
\value{
A tibble.
}
\description{
Sets the necessary data frame for interactive analysis plotting using
Python's plotly library.
}
