% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{sfilter,analysis-method}
\alias{sfilter,analysis-method}
\title{Filters an analysis object so as to keep only those
anomalies whose strength is greater than x.}
\usage{
\S4method{sfilter}{analysis}(object, x, f = minmax_normalization)
}
\arguments{
\item{object}{An analysis object}

\item{x}{The strength threshold}

\item{f}{A function used to standardize strengths (otherwise collective and point
anomalies are incomparable). Defaults to min-max normalization.}
}
\value{
An analysis object.
}
\description{
Filters an analysis object so as to keep only those
anomalies whose strength is greater than x.
}
