% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{show,analysis-method}
\alias{show,analysis-method}
\title{Show method for the EEG class that calls View on the canoms panoms
and eeg slots.}
\usage{
\S4method{show}{analysis}(object)
}
\arguments{
\item{object}{An Analysis object.}
}
\description{
Show method for the EEG class that calls View on the canoms panoms
and eeg slots.
}
