% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{show,eeg-method}
\alias{show,eeg-method}
\title{Show method for the EEG class that prints the data slot.}
\usage{
\S4method{show}{eeg}(object)
}
\arguments{
\item{object}{An EEG object.}
}
\description{
Show method for the EEG class that prints the data slot.
}
