% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{sigma_index}
\alias{sigma_index}
\title{Sigma index method as described in O'Reilly & Nielsen (2015) and
Hupponen et al. (2007).
Parameters:
x: numeric vector
fs = 500: sampling frequency}
\usage{
sigma_index(x, fs = 500)
}
\description{
Returns sigma index of vector x of EEG samples.
}
