% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{spectrogram,eeg-method}
\alias{spectrogram,eeg-method}
\title{Produces the spectrogram of an EEG object.
Merele a wrapper for rsleep's spectrogram}
\usage{
\S4method{spectrogram}{eeg}(object, channel, maxFreq = 30, freq = 4)
}
\arguments{
\item{object}{The EEG object to be analyzed.}

\item{channel}{The channel to be analyzed.}
}
\description{
Produces the spectrogram of an EEG object.
Merele a wrapper for rsleep's spectrogram
}
