% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{standardize_strengths,analysis-method}
\alias{standardize_strengths,analysis-method}
\title{Normalizes point and collective anomalies' strengths
given a normalizing function f.}
\usage{
\S4method{standardize_strengths}{analysis}(object, f)
}
\value{
An analysis object.
}
\description{
Normalizes point and collective anomalies' strengths
given a normalizing function f.
}
