% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifactor.R
\name{stepwise_analysis}
\alias{stepwise_analysis}
\title{Performs stepwise (or epoch by epoch) CAPA analysis on EEG data.}
\usage{
stepwise_analysis(
  eeg,
  step_size,
  res = 1,
  alpha = 8,
  beta = 1,
  thresh = 3,
  write = FALSE
)
}
\arguments{
\item{eeg}{An eeg object.}

\item{step_size}{Size in seconds of each step.}

\item{res}{Resolution at which to perform the analyses. Defaults to 1.}

\item{alpha}{Threshold of strength significance for collective anomalies.}

\item{beta}{Threshold of strength significance for point anomalies.}

\item{thresh}{How many seconds anomaly n must be from anomaly (n - 1) to
take them both as a single anomaly?}
}
\value{
An analysis object.
}
\description{
Performs stepwise (or epoch by epoch) CAPA analysis on EEG data.
}
