% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{subset_eeg,eeg-method}
\alias{subset_eeg,eeg-method}
\title{Subsets an EEG object from second s to e and
returns new EEG object whose data is that subset.}
\usage{
\S4method{subset_eeg}{eeg}(object, s, e)
}
\arguments{
\item{object}{An eeg object.}

\item{s}{Starting time of the subset in seconds.}

\item{e}{Ending time of the subset in seconds.}
}
\value{
A new eeg whose data is the subset ranging from
second s to e of the object's data attribute.
}
\description{
Subsets an EEG object from second s to e and
returns new EEG object whose data is that subset.
}
