% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cln.R
\name{thresh}
\alias{thresh}
\title{Given an collective anomalies (or canoms) data frame,
an index  i and a threshold in seconds, returns the time at which}
\usage{
thresh(canoms, i, threshold)
}
\arguments{
\item{canoms}{Collective anomalies data frame, such as the @canoms attribute in an analysis object.}

\item{i}{Anomaly index}

\item{threshold}{Threshold in seconds}
}
\value{
An integer. The time of occurence of anomaly canoms[i] plus threshold.
}
\description{
Given an collective anomalies (or canoms) data frame,
an index  i and a threshold in seconds, returns the time at which
}
