% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{update.epochs}
\alias{update.epochs}
\title{Update a data frame containing anomalous epoch-subepoch
pairs given a certain analysis results.}
\usage{
\method{update}{epochs}(epoch_data, analysis)
}
\value{
A data frame as defined by create.epoch.data().

An analysis object.
}
\description{
Update a data frame containing anomalous epoch-subepoch
pairs given a certain analysis results.
}
