% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dman.R
\name{update_epochs}
\alias{update_epochs}
\title{Update a data frame containing anomalous epoch-subepoch
pairs given a certain analysis results.}
\usage{
update_epochs(epoch_data, analysis)
}
\value{
A data frame as defined by create_epoch_data().

An analysis object.
}
\description{
Update a data frame containing anomalous epoch-subepoch
pairs given a certain analysis results.
}
