% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{vbandpass}
\alias{vbandpass}
\title{Wrapper function for bandpass filtering
a vector given filtering frequencies and
a sampling frequency.}
\usage{
vbandpass(vec, l, h, fs)
}
\description{
Wrapper function for bandpass filtering
a vector given filtering frequencies and
a sampling frequency.
}
