% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{vhigh_pass}
\alias{vhigh_pass}
\title{Wrapper function for high-pass filtering
a vector given a filtering frequency and
a sampling frequency.}
\usage{
vhigh_pass(vec, n, fs)
}
\description{
Wrapper function for high-pass filtering
a vector given a filtering frequency and
a sampling frequency.
}
