% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{vlow_pass}
\alias{vlow_pass}
\title{Wrapper function for low-pass filtering
a vector given a filtering frequency and
a sampling frequency.}
\usage{
vlow_pass(vec, n, fs)
}
\description{
Wrapper function for low-pass filtering
a vector given a filtering frequency and
a sampling frequency.
}
