% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{drop_epochs,eeg-method}
\alias{drop_epochs,eeg-method}
\title{Given an eeg object and a list of epochs,
removes all samples from the given epochs.}
\usage{
\S4method{drop_epochs}{eeg}(object, epochs, epoch = 30)
}
\arguments{
\item{object}{An eeg object.}

\item{epochs}{A vector of natural integers.}
}
\value{
An EEG object.
}
\description{
Notice that the extremes of the epoch are kept.
}
