% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{class}
\name{EEG}
\alias{EEG}
\title{Class providing object with methods for communication with lightning-viz server}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} with methods for EEG analysis,
visualization and manipulation.
}
\description{
Class providing object with methods for communication with lightning-viz server

Class providing object with methods for communication with lightning-viz server
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Stores de EEG data in a data frame format (tibble by default).}

\item{\code{signals}}{Stores a data frame with information on the EEG signals, if a
`signals` file was provided on initialization.}

\item{\code{canoms}}{A data frame with all data pertaining to collective anomalies
(artifacts) found on the EEG. Empty by default and until artifact analysis is
performed.}

\item{\code{panoms}}{A data frame with all data pertaining to point anomalies
(artifacts) found on the EEG. Empty by default and until artifact analysis is
performed.}

\item{\code{psd}}{Power spectrum density data of this EEG. Empty by default and
until PSD is computed on the EEG.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EEG-new}{\code{EEG$new()}}
\item \href{#method-EEG-subset_by_seconds}{\code{EEG$subset_by_seconds()}}
\item \href{#method-EEG-subset}{\code{EEG$subset()}}
\item \href{#method-EEG-resample}{\code{EEG$resample()}}
\item \href{#method-EEG-get_fs}{\code{EEG$get_fs()}}
\item \href{#method-EEG-vlow_pass}{\code{EEG$vlow_pass()}}
\item \href{#method-EEG-vhigh_pass}{\code{EEG$vhigh_pass()}}
\item \href{#method-EEG-vbandpass}{\code{EEG$vbandpass()}}
\item \href{#method-EEG-low_pass}{\code{EEG$low_pass()}}
\item \href{#method-EEG-high_pass}{\code{EEG$high_pass()}}
\item \href{#method-EEG-bandpass}{\code{EEG$bandpass()}}
\item \href{#method-EEG-plot_channel}{\code{EEG$plot_channel()}}
\item \href{#method-EEG-plot}{\code{EEG$plot()}}
\item \href{#method-EEG-iplot}{\code{EEG$iplot()}}
\item \href{#method-EEG-drop_epochs}{\code{EEG$drop_epochs()}}
\item \href{#method-EEG-drop_subepochs}{\code{EEG$drop_subepochs()}}
\item \href{#method-EEG-artf}{\code{EEG$artf()}}
\item \href{#method-EEG-artf_stepwise}{\code{EEG$artf_stepwise()}}
\item \href{#method-EEG-get_contaminated_channels}{\code{EEG$get_contaminated_channels()}}
\item \href{#method-EEG-set_plot_data}{\code{EEG$set_plot_data()}}
\item \href{#method-EEG-plot_channel_artifacts}{\code{EEG$plot_channel_artifacts()}}
\item \href{#method-EEG-plot_artifacts}{\code{EEG$plot_artifacts()}}
\item \href{#method-EEG-sfilter}{\code{EEG$sfilter()}}
\item \href{#method-EEG-get_contaminated_epochs}{\code{EEG$get_contaminated_epochs()}}
\item \href{#method-EEG-artf_reject}{\code{EEG$artf_reject()}}
\item \href{#method-EEG-spectrogram}{\code{EEG$spectrogram()}}
\item \href{#method-EEG-channel_psd}{\code{EEG$channel_psd()}}
\item \href{#method-EEG-compute_psd}{\code{EEG$compute_psd()}}
\item \href{#method-EEG-plot_psd}{\code{EEG$plot_psd()}}
\item \href{#method-EEG-iplot_psd}{\code{EEG$iplot_psd()}}
\item \href{#method-EEG-clone}{\code{EEG$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-new"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$new(data_file, signals_file = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-subset_by_seconds"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-subset_by_seconds}{}}}
\subsection{Method \code{subset_by_seconds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$subset_by_seconds(s, e)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-subset"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-subset}{}}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$subset(s, e, epoch = 30)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-resample"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-resample}{}}}
\subsection{Method \code{resample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$resample(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-get_fs"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-get_fs}{}}}
\subsection{Method \code{get_fs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$get_fs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-vlow_pass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-vlow_pass}{}}}
\subsection{Method \code{vlow_pass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$vlow_pass(vec, n, fs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-vhigh_pass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-vhigh_pass}{}}}
\subsection{Method \code{vhigh_pass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$vhigh_pass(vec, n, fs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-vbandpass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-vbandpass}{}}}
\subsection{Method \code{vbandpass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$vbandpass(vec, l, h, fs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-low_pass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-low_pass}{}}}
\subsection{Method \code{low_pass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$low_pass(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Filter frequency in Hz
Given a numeric frequency n,
applies a high-pass Butterworth filter to all channels.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-high_pass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-high_pass}{}}}
\subsection{Method \code{high_pass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$high_pass(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Filter frequency in Hz

Given a numeric frequency n,
applies a bandpass filter to all channels.}

\item{\code{object}}{An eeg object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-bandpass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-bandpass}{}}}
\subsection{Method \code{bandpass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$bandpass(l, h)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{l}}{lower bound in Hz}

\item{\code{h}}{higher bound in Hz

Given a integer n, plots EEG record of the nth channel.}

\item{\code{object}}{An eeg object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-plot_channel"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-plot_channel}{}}}
\subsection{Method \code{plot_channel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$plot_channel(channel)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{channel}}{An integer indicating index of channel to plot.

Plot all the EEG channels in a vertical single-column layout.
Plot an interactive plotly figure of all EEG channels.
Given a list of epochs, remove those epochs from the EEG data.
Note that the extremes of the epoch are kept.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-plot"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-iplot"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-iplot}{}}}
\subsection{Method \code{iplot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$iplot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-drop_epochs"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-drop_epochs}{}}}
\subsection{Method \code{drop_epochs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$drop_epochs(epochs, epoch = 30)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{epochs}}{A vector of natural integers.
Given  a list of subepochs, removes epoch-subepoch                    
pairs from the eeg data.

Note that elements in the epoch and subepoch lists are assumed to have
 an element-wise association. This means that, if the lists
are (e_1, ..., e_n) and (s_1, ..., s_n), then epoch/subepoch pairs
                         (e_i, s_i)
are removed, for i in [1, n].}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-drop_subepochs"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-drop_subepochs}{}}}
\subsection{Method \code{drop_subepochs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$drop_subepochs(epochs, subepochs, epoch = 30)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{epochs}}{A list of integers.}

\item{\code{subepochs}}{A list of integers.}

\item{\code{epoch}}{An integer: how many seconds make up an epoch?

Given an EEG channel, returns a data frame
whose format is idoneous for plotting the analysis results.
Such format consists of the EEG Time measure (x axis), the raw
EEG data of the chanel and a column that is filled with NaN except
at each time instance where any type of anomaly was found.
This method is not intended for isolated calls, but is implicitely used
in other plotting methods.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-artf"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-artf}{}}}
\subsection{Method \code{artf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$artf(alpha = 8, beta = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-artf_stepwise"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-artf_stepwise}{}}}
\subsection{Method \code{artf_stepwise()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$artf_stepwise(step_size = 30, alpha = 8)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-get_contaminated_channels"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-get_contaminated_channels}{}}}
\subsection{Method \code{get_contaminated_channels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$get_contaminated_channels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-set_plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-set_plot_data}{}}}
\subsection{Method \code{set_plot_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$set_plot_data(chan)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chan}}{An integer that points to the EEG channel.

Plots analysis results for a specific channel.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-plot_channel_artifacts"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-plot_channel_artifacts}{}}}
\subsection{Method \code{plot_channel_artifacts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$plot_channel_artifacts(chan, size = 0.2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chan}}{An integer that points to the EEG channel.}

\item{\code{size}}{(optional) The size of the red dots that signal an anomaly.

Plots all channels with any kind of anomalies.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-plot_artifacts"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-plot_artifacts}{}}}
\subsection{Method \code{plot_artifacts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$plot_artifacts(size = 0.2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{(optional)}}{size Size of the red dots that depict anomalies.

Filters the EEG so as to keep only those
anomalies whose normalized strength is greater than x.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-sfilter"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-sfilter}{}}}
\subsection{Method \code{sfilter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$sfilter(x, f = minmax_normalization)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The strength threshold}

\item{\code{f}}{A function used to normalize strengths (otherwise collective
and point anomalies are incomparable). Defaults to min-max normalization.

Helper function

Computes a data frame with the epoch-subepoch pairs
found to contain anomalies, and their respective average strengths.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-get_contaminated_epochs"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-get_contaminated_epochs}{}}}
\subsection{Method \code{get_contaminated_epochs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$get_contaminated_epochs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame containing epoch-subepoch pairs
found to contain artifacts as well as the average
strength of the artifacts for each subepoch.

Returns an artifact-rejected copy of this EEG.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-artf_reject"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-artf_reject}{}}}
\subsection{Method \code{artf_reject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$artf_reject()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An EEG object.

Plot the spectrum of all channels.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-spectrogram"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-spectrogram}{}}}
\subsection{Method \code{spectrogram()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$spectrogram(channel, max_freq = 30, freq = 4)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-channel_psd"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-channel_psd}{}}}
\subsection{Method \code{channel_psd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$channel_psd(channel)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-compute_psd"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-compute_psd}{}}}
\subsection{Method \code{compute_psd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$compute_psd()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-plot_psd"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-plot_psd}{}}}
\subsection{Method \code{plot_psd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$plot_psd(xlim = 250)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{(optional)}}{xlim An integer, positive bound for the x-axis.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-iplot_psd"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-iplot_psd}{}}}
\subsection{Method \code{iplot_psd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$iplot_psd(xlim = 250)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
