% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{EEG}
\alias{EEG}
\title{EEG}
\description{
Abstract base class for EEG data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{(\code{tibble})\cr
Data frame (tibble) with the EEG data.}

\item{\code{canoms}}{(\code{tibble})\cr
Data frame (tibble) with the collective anomalies data.}

\item{\code{panoms}}{(\code{tibble})\cr
Data frame (tibble) with the point anomalies data.}

\item{\code{psd}}{(\code{tibble})\cr
Data frame (tibble) with the power spectral density data.}

\item{\code{spindles}}{(\code{tibble})\cr
Data frame (tibble) with the spindle data.}

\item{\code{fs}}{(\code{numeric})
Sampling frequency of the EEG.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EEG-new}{\code{EEG$new()}}
\item \href{#method-EEG-subset_by_seconds}{\code{EEG$subset_by_seconds()}}
\item \href{#method-EEG-subset}{\code{EEG$subset()}}
\item \href{#method-EEG-resample}{\code{EEG$resample()}}
\item \href{#method-EEG-get_fs}{\code{EEG$get_fs()}}
\item \href{#method-EEG-vlow_pass}{\code{EEG$vlow_pass()}}
\item \href{#method-EEG-vhigh_pass}{\code{EEG$vhigh_pass()}}
\item \href{#method-EEG-vbandpass}{\code{EEG$vbandpass()}}
\item \href{#method-EEG-low_pass}{\code{EEG$low_pass()}}
\item \href{#method-EEG-high_pass}{\code{EEG$high_pass()}}
\item \href{#method-EEG-bandpass}{\code{EEG$bandpass()}}
\item \href{#method-EEG-plot_channel}{\code{EEG$plot_channel()}}
\item \href{#method-EEG-plot}{\code{EEG$plot()}}
\item \href{#method-EEG-iplot}{\code{EEG$iplot()}}
\item \href{#method-EEG-drop_epochs}{\code{EEG$drop_epochs()}}
\item \href{#method-EEG-drop_subepochs}{\code{EEG$drop_subepochs()}}
\item \href{#method-EEG-artf}{\code{EEG$artf()}}
\item \href{#method-EEG-artf_stepwise}{\code{EEG$artf_stepwise()}}
\item \href{#method-EEG-set_anom_dfs}{\code{EEG$set_anom_dfs()}}
\item \href{#method-EEG-set_anom_time_features}{\code{EEG$set_anom_time_features()}}
\item \href{#method-EEG-get_contaminated_channels}{\code{EEG$get_contaminated_channels()}}
\item \href{#method-EEG-set_plot_data}{\code{EEG$set_plot_data()}}
\item \href{#method-EEG-plot_channel_artifacts}{\code{EEG$plot_channel_artifacts()}}
\item \href{#method-EEG-plot_artifacts}{\code{EEG$plot_artifacts()}}
\item \href{#method-EEG-gen_artifact_plots}{\code{EEG$gen_artifact_plots()}}
\item \href{#method-EEG-has_anoms}{\code{EEG$has_anoms()}}
\item \href{#method-EEG-sfilter}{\code{EEG$sfilter()}}
\item \href{#method-EEG-get_contaminated_epochs}{\code{EEG$get_contaminated_epochs()}}
\item \href{#method-EEG-artf_reject}{\code{EEG$artf_reject()}}
\item \href{#method-EEG-spectrogram}{\code{EEG$spectrogram()}}
\item \href{#method-EEG-channel_psd}{\code{EEG$channel_psd()}}
\item \href{#method-EEG-compute_psd}{\code{EEG$compute_psd()}}
\item \href{#method-EEG-plot_psd}{\code{EEG$plot_psd()}}
\item \href{#method-EEG-iplot_psd}{\code{EEG$iplot_psd()}}
\item \href{#method-EEG-spindle_detection}{\code{EEG$spindle_detection()}}
\item \href{#method-EEG-plot_spindle_distribution}{\code{EEG$plot_spindle_distribution()}}
\item \href{#method-EEG-clone}{\code{EEG$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-new"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the EEG class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$new(data_file, set_epochs = TRUE, epoch = 30)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{(\code{string})
Directory of .csv file containing EEG data.}

\item{\code{set_epochs}}{(\code{bool = TRUE})
Whether to set the epoch and subepoch features upon reading the
EEG data. Defaults to TRUE, but should be FALSE if the EEG data
being read already has these features.}

\item{\code{epoch}}{(\code{int})
How many seconds make up an epoch? Only relevant if set_epochs is TRUE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-subset_by_seconds"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-subset_by_seconds}{}}}
\subsection{Method \code{subset_by_seconds()}}{
Subsets EEG data to keep only measures in the interval \link{s, e}, where
\code{s} and \code{e} are times in seconds and \code{e > s}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$subset_by_seconds(s, e)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{(\code{integer}).}

\item{\code{e}}{(\code{integer}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-subset"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-subset}{}}}
\subsection{Method \code{subset()}}{
Subsets EEG data to keep only measures in the interval \link{s, e}, where
\code{s} and \code{e} are epochs and \code{e >= s}. If \code{e == s} then only data from
that single epoch is kept.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$subset(s, e)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{(\code{integer}).}

\item{\code{e}}{(\code{integer}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-resample"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-resample}{}}}
\subsection{Method \code{resample()}}{
Subsets EEG data to keep only one for every \code{n} samples. This is a
brute resampling algorithm and should only not be used except for
special purposes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$resample(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{(\code{integer}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-get_fs"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-get_fs}{}}}
\subsection{Method \code{get_fs()}}{
Returns the sampling frequency of the EEG data. The frequency is computed
from the data and not extracted from the $fs field. This method is useful
for updating the $fs field after resampling or other data modifications.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$get_fs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-vlow_pass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-vlow_pass}{}}}
\subsection{Method \code{vlow_pass()}}{
Private (do not use)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$vlow_pass(vec, n, fs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vec}}{(\code{vector}).}

\item{\code{n}}{(\code{int}).}

\item{\code{fs}}{(\code{int}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-vhigh_pass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-vhigh_pass}{}}}
\subsection{Method \code{vhigh_pass()}}{
Private (do not use)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$vhigh_pass(vec, n, fs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vec}}{(\code{vector}).}

\item{\code{n}}{(\code{int}).}

\item{\code{fs}}{(\code{int}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-vbandpass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-vbandpass}{}}}
\subsection{Method \code{vbandpass()}}{
Private (do not use)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$vbandpass(vec, l, h, fs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vec}}{(\code{vector}).}

\item{\code{l}}{(\code{int}).}

\item{\code{h}}{(\code{int}).}

\item{\code{fs}}{(\code{int}).}

\item{\code{n}}{(\code{int}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-low_pass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-low_pass}{}}}
\subsection{Method \code{low_pass()}}{
Applies a low-pass filter to the EEG data. The value of \code{n} is used
to compute the pass filter \code{F = n / (fs / 2)} (Nyquist frequency).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$low_pass(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{(\code{numeric}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-high_pass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-high_pass}{}}}
\subsection{Method \code{high_pass()}}{
Applies a high-pass filter to the EEG data. The value of \code{n} is used
to compute the pass filter \code{F = n / (fs / 2)} (Nyquist frequency).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$high_pass(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{(\code{numeric}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-bandpass"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-bandpass}{}}}
\subsection{Method \code{bandpass()}}{
Applies a bandpass filter to the EEG data. Parameters \code{l} and \code{h} are
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$bandpass(l, h)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{l}}{(\code{numeric}).}

\item{\code{h}}{(\code{numeric}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-plot_channel"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-plot_channel}{}}}
\subsection{Method \code{plot_channel()}}{
Plots the EEG channel number \code{channel}. If \code{s} and \code{e} are non-zero the
plot only shows EEG data from epoch \code{s} to epoch \code{e} (inclusive). If
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$plot_channel(channel, s = 0, e = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{channel}}{(\code{integer}).}

\item{\code{s}}{(\code{integer}) = 0}

\item{\code{e}}{(\code{integer}) = 0}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-plot"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the EEG data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
cowplot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-iplot"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-iplot}{}}}
\subsection{Method \code{iplot()}}{
Produces an interactive PlotLy plot of the EEG data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$iplot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
plotly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-drop_epochs"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-drop_epochs}{}}}
\subsection{Method \code{drop_epochs()}}{
Removes from the EEG data the samples of any epoch in \code{epochs}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$drop_epochs(epochs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{epochs}}{(\code{int})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-drop_subepochs"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-drop_subepochs}{}}}
\subsection{Method \code{drop_subepochs()}}{
Removes from the EEG data the samples of any epoch-subepoch pair
in \code{epochs}, \code{subepochs}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$drop_subepochs(epochs, subepochs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{epochs}}{(\code{int})}

\item{\code{subepochs}}{(\code{int})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-artf"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-artf}{}}}
\subsection{Method \code{artf()}}{
Performs MV-CAPA analysis on the EEG data; formats the results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$artf(alpha = 0.1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{(\code{numeric})\cr
Only collective anomalies with strength greater than \code{alpha} are kept.
Defaults to \code{8}.}

\item{\code{beta}}{(\code{numeric})\cr
Only point anomalies with strength greater than \code{beta} are kept.
Defaults to \code{1}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-artf_stepwise"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-artf_stepwise}{}}}
\subsection{Method \code{artf_stepwise()}}{
Performs MV-CAPA analysis on the EEG data by steps of size \code{step_size}
(in seconds); formats the results and sets them to the \code{canoms} and
\code{panoms} fields.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$artf_stepwise(step_size = 30, alpha = 0.05, type = "mean", verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step_size}}{(\code{int}) Seconds per analyzed step.}

\item{\code{alpha}}{(\code{numeric}) The result will exclude anoms with normalized strength <= alpha.
The value of \code{alpha} must satisfie 0 ≤ alpha ≤ 1.
Defaults to \code{0.05}.}

\item{\code{type}}{(\code{string}) Should be "mean" or "meanvar", depending on
whether to test for changes in mean or in mean and variance.}

\item{\code{verbose}}{(\code{bool}) Print log messages of the analysis process?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-set_anom_dfs"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-set_anom_dfs}{}}}
\subsection{Method \code{set_anom_dfs()}}{
A helper function to be privately used in artf_stepwise. Transforms each
tibble with the joined data of each analysis; it adjusts the location
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$set_anom_dfs(mps, panoms, canoms, alpha = 0.05)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mps}}{(\code{vector}) A vector containing the number of measures per analyzed step of data.}

\item{\code{panoms}}{(\code{list}) List of data frames, each comming from calling
anomaly::point_anomalies on an analyzed subset of the EEG data.}

\item{\code{canoms}}{(\code{numeric}) List of data frames, each comming from calling
anomaly::collective_anomalies on an analyzed subset of the EEG data.}

\item{\code{alpha}}{See artf_stepwise which transfers its alpha parameter to this.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-set_anom_time_features"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-set_anom_time_features}{}}}
\subsection{Method \code{set_anom_time_features()}}{
Private wrapper.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$set_anom_time_features()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-get_contaminated_channels"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-get_contaminated_channels}{}}}
\subsection{Method \code{get_contaminated_channels()}}{
Private wrapper.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$get_contaminated_channels()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-set_plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-set_plot_data}{}}}
\subsection{Method \code{set_plot_data()}}{
(Private)
Returns a df whose format is idoneous for plotting the analysis results.
Such format consists of the EEG Time measure (x axis), the raw
EEG data of the chanel and a column that is filled with NaN except
at each time instance where any type of anomaly was found.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$set_plot_data(chan, s = 0, e = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chan}}{(\code{int}) Integer in the range \link{1, number_of_eeg_signals}.}

\item{\code{s}}{Plot should begin at epoch s.}

\item{\code{e}}{Plot should end at epoch e.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-plot_channel_artifacts"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-plot_channel_artifacts}{}}}
\subsection{Method \code{plot_channel_artifacts()}}{
Plots analysis results for a specific channel.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$plot_channel_artifacts(chan, s = 0, e = 0, size = 0.2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chan}}{An integer that points to the EEG channel.}

\item{\code{s}}{Plot should begin at epoch s.}

\item{\code{e}}{Plot should end at epoch e.}

\item{\code{size}}{(optional) The size of the red dots that signal an anomaly.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-plot_artifacts"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-plot_artifacts}{}}}
\subsection{Method \code{plot_artifacts()}}{
Plots all channels with any kind of anomalies from epoch \code{s} to \code{e}.
Mark anomalies with dots of radius \code{size}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$plot_artifacts(s = 0, e = 0, size = 0.2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{Starting epoch}

\item{\code{e}}{Ending epoch}

\item{\code{size}}{Size of the red dots that mark anomalies in the plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
cowplot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-gen_artifact_plots"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-gen_artifact_plots}{}}}
\subsection{Method \code{gen_artifact_plots()}}{
Saves to \code{dir} a plot of all artifacts. Each plot contains a number of
of epochs determined by \code{epochs_per_plot}. If \code{ignore_clean} is \code{TRUE},
then any segment not containing anomalies is not plotted.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$gen_artifact_plots(epochs_per_plot, dir = getwd(), ignore_clean = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{epochs_per_plot}}{Size of the x-axis in epochs.}

\item{\code{dir}}{Directory where plots will be saved.}

\item{\code{ignore_clean}}{Skip plotting clean segments?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-has_anoms"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-has_anoms}{}}}
\subsection{Method \code{has_anoms()}}{
Determines whether a data segment contains anomalies. If \code{e} parameter
is zero (default), the function test whether any anomaly exists in
the whole EEG data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$has_anoms(s, e = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{Starting epoch of the segment}

\item{\code{e}}{Ending epoch of the segment, defaults to zero.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
bool
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-sfilter"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-sfilter}{}}}
\subsection{Method \code{sfilter()}}{
Normalizes the strengths of point and collective anomalies
given a normalizing function \code{f}, which defaults to min-max
normalization, and removes anomalies whose strength is inferior
to \code{x}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$sfilter(x, f = minmax_normalization)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Threshold value}

\item{\code{f}}{Normalization function (defaults to min-max).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-get_contaminated_epochs"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-get_contaminated_epochs}{}}}
\subsection{Method \code{get_contaminated_epochs()}}{
COMPLETE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$get_contaminated_epochs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-artf_reject"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-artf_reject}{}}}
\subsection{Method \code{artf_reject()}}{
Performs artifact rejection by means of dropping from the EEG data
all segments corresponding to epoch-subepoch pairs
that were found to contain anomalies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$artf_reject()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-spectrogram"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-spectrogram}{}}}
\subsection{Method \code{spectrogram()}}{
Plots the spectrogram of an EEG signal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$spectrogram(channel, max_freq = 30, freq = 4)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{channel}}{Integer pointing to an EEG signal.}

\item{\code{max_freq}}{Maximum frequency of the plot.}

\item{\code{freq}}{Aggregate frequency to set plot resolution.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-channel_psd"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-channel_psd}{}}}
\subsection{Method \code{channel_psd()}}{
Computes the PSD of an EEG signal. It is returned in data frame format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$channel_psd(channel)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{channel}}{Integer pointing to an EEG signal.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-compute_psd"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-compute_psd}{}}}
\subsection{Method \code{compute_psd()}}{
Computes the PSD of each EEG signal. Result is set to the $psd field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$compute_psd()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-plot_psd"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-plot_psd}{}}}
\subsection{Method \code{plot_psd()}}{
Produces a ggplot of the PSD. \code{compute_psd} must have been called before.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$plot_psd(xlim = 250)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xlim}}{Maximum frequency to show}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-iplot_psd"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-iplot_psd}{}}}
\subsection{Method \code{iplot_psd()}}{
Produces a plotly plot of the PSD. \code{compute_psd} must have been called before.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$iplot_psd(xlim = 250)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xlim}}{Maximum frequency to show}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly figure
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-spindle_detection"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-spindle_detection}{}}}
\subsection{Method \code{spindle_detection()}}{
Performs spindle detection either on a specific signal (if \code{channel} is
non-zero) or for each EEG signal. The two available methods are
Sigma Index and Relative Spindle Power. For details on these, consult the
GitHub README. The result is set to the $spindles field.
'
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$spindle_detection(channel = 0, method = "sigma_index", filter = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{channel}}{On which signal to detect spindles? If zero (default),
all signals are subjected to spindle detection.}

\item{\code{method}}{Either "sigma_index" (default) or "rsp".}

\item{\code{filter}}{A boolean determining whether to filter spindles according
to standard strength thresholds.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-plot_spindle_distribution"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-plot_spindle_distribution}{}}}
\subsection{Method \code{plot_spindle_distribution()}}{
Plots the distribution of spindles across the EEG. \code{spindle_detection}
must have been previously called.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$plot_spindle_distribution(
  channel = 0,
  time_axis = "epoch",
  xbins = 10,
  ybins = 10,
  from = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{channel}}{On which signal to detect spindles? If zero (default),
a cumulative index of the spindles on all signals is
plotted.}

\item{\code{time_axis}}{Show distribution over "epoch", "second", "minute" or "hour"?}

\item{\code{xbins}}{Size of the x-bins in the plot.}

\item{\code{ybins}}{Size of the y-bins in the plot.}

\item{\code{from}}{Start of the plot. Defaults to 0 (whole EEG).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
void
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EEG-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EEG-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EEG$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
