% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{get_channel_psd,eeg-method}
\alias{get_channel_psd,eeg-method}
\title{Given an eeg object and a channel's column index,
compute the power spectrum density of  the channel
using Welch's method.}
\usage{
\S4method{get_channel_psd}{eeg}(object, channel)
}
\arguments{
\item{object}{An eeg object.}

\item{channel}{A column number}
}
\value{
An data frame containing the power density
for every frequency composing the given channel's signal.
}
\description{
Given an eeg object and a channel's column index,
compute the power spectrum density of  the channel
using Welch's method.
}
