% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg.R
\name{get_sampling_frequency,eeg-method}
\alias{get_sampling_frequency,eeg-method}
\title{Returns sampling frequency}
\usage{
\S4method{get_sampling_frequency}{eeg}(object)
}
\arguments{
\item{object}{An eeg object.}
}
\value{
An number, the sampling frequency
}
\description{
Returns sampling frequency
}
