% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{psd,eeg-method}
\alias{psd,eeg-method}
\title{Given an eeg object, compute the average power spectrum density
of each channel across epochs.}
\usage{
\S4method{psd}{eeg}(object, epoch = 30)
}
\arguments{
\item{object}{An eeg object.}
}
\value{
A data frame with the spectrum density for every frequency
for every channel in the  EEG record.
}
\description{
Given an eeg object, compute the average power spectrum density
of each channel across epochs.
}
