% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{spectogram,eeg-method}
\alias{spectogram,eeg-method}
\title{Produces the spectogram of an EEG object.}
\usage{
\S4method{spectogram}{eeg}(object, channel, hcolors = 10)
}
\arguments{
\item{object}{The EEG object to be analyzed.}

\item{channel}{The channel to be analyzed.}
}
\description{
Produces the spectogram of an EEG object.
}
